requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
pandas==2.1.3
requests==2.31.0
python-dotenv==1.0.0
PyInstaller==6.16.0
SQLAlchemy==2.0.23
WTForms==3.1.0
